requirements.txt
cssselect
lxml
requests
python-decouple
